Cleaned project and added readme, ci workflows
