+++ .github/workflows/build.yml
@@ -1,0 +1,16 @@
+name: Gradle Build
+on: [push, pull_request]
+
+jobs:
+  check:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up JDK 11
+        uses: actions/setup-java@v2
+        with:
+          java-version: 11
+          distribution: adopt
+      - name: Run Gradle Check
+        run: |
+          ./gradlew check
