+++ .github/workflows/wrapper.yml
@@ -1,0 +1,10 @@
+name: Validate Gradle Wrapper
+on: [push, pull_request]
+
+jobs:
+  validate:
+    name: Validate
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: gradle/wrapper-validation-action@v1
