+++ src/main/java/transportservice/RunPlugin.java
@@ -4,14 +4,10 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.Version;
 import org.opensearch.cluster.ClusterModule;
-import org.opensearch.cluster.node.DiscoveryNode;
-import org.opensearch.common.SuppressForbidden;
 import org.opensearch.common.io.stream.NamedWriteableRegistry;
 import org.opensearch.common.network.NetworkModule;
 import org.opensearch.common.network.NetworkService;
 import org.opensearch.common.settings.Settings;
-import org.opensearch.common.transport.TransportAddress;
-import org.opensearch.common.unit.TimeValue;
 import org.opensearch.common.util.PageCacheRecycler;
 import org.opensearch.indices.IndicesModule;
 import org.opensearch.indices.breaker.CircuitBreakerService;
@@ -21,8 +17,6 @@ import org.opensearch.threadpool.ThreadPool;
 import org.opensearch.transport.*;
 import transportservice.netty4.Netty4Transport;
 
-import java.io.IOException;
-import java.io.UncheckedIOException;
 import java.net.*;
 import java.util.Collections;
 import java.util.List;
@@ -30,7 +24,6 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static java.util.Collections.emptyMap;
 import static java.util.Collections.emptySet;
 
 public class RunPlugin {
@@ -38,37 +31,28 @@ public class RunPlugin {
     private static final Settings settings = Settings.builder()
             .put("node.name", "NettySizeHeaderFrameDecoderTests")
             .put(TransportSettings.BIND_HOST.getKey(), "127.0.0.1")
-            .put(TransportSettings.PORT.getKey(), "0")
+            .put(TransportSettings.PORT.getKey(), "9301")
             .build();
     private static final Logger logger = LogManager.getLogger(RunPlugin.class);
     public static final TransportInterceptor NOOP_TRANSPORT_INTERCEPTOR = new TransportInterceptor() {
     };
-    private static final Version CURRENT_VERSION = Version.fromString(String.valueOf(Version.CURRENT.major) + ".0.0");
-    protected static final Version version0 = CURRENT_VERSION.minimumCompatibilityVersion();
+    ThreadPool threadPool = new ThreadPool(settings);
+    NetworkService networkService = new NetworkService(Collections.emptyList());
+    PageCacheRecycler pageCacheRecycler = new PageCacheRecycler(settings);
+    IndicesModule indicesModule = new IndicesModule(Collections.emptyList());
+    SearchModule searchModule = new SearchModule(settings, Collections.emptyList());
+    List<NamedWriteableRegistry.Entry> namedWriteables = Stream.of(
+            NetworkModule.getNamedWriteables().stream(),
+            indicesModule.getNamedWriteables().stream(),
+            searchModule.getNamedWriteables().stream(),
+            null,
+            ClusterModule.getNamedWriteables().stream()
+    ).flatMap(Function.identity()).collect(Collectors.toList());
+    final NamedWriteableRegistry namedWriteableRegistry = new NamedWriteableRegistry(namedWriteables);
+    final CircuitBreakerService circuitBreakerService = new NoneCircuitBreakerService();
 
 
-    @SuppressForbidden(reason = "need local ephemeral port")
-    protected static InetSocketAddress getLocalEphemeral() throws UnknownHostException {
-        return new InetSocketAddress(InetAddress.getLocalHost(), 0);
-    }
-
-    public static void main(String[] args) throws UnknownHostException {
-        ThreadPool threadPool = new ThreadPool(settings);
-        NetworkService networkService = new NetworkService(Collections.emptyList());
-        PageCacheRecycler pageCacheRecycler = new PageCacheRecycler(settings);
-        IndicesModule indicesModule = new IndicesModule(Collections.emptyList());
-        SearchModule searchModule = new SearchModule(settings, Collections.emptyList());
-        List<NamedWriteableRegistry.Entry> namedWriteables = Stream.of(
-                NetworkModule.getNamedWriteables().stream(),
-                indicesModule.getNamedWriteables().stream(),
-                searchModule.getNamedWriteables().stream(),
-                null,
-                ClusterModule.getNamedWriteables().stream()
-        ).flatMap(Function.identity()).collect(Collectors.toList());
-        final NamedWriteableRegistry namedWriteableRegistry = new NamedWriteableRegistry(namedWriteables);
-        final CircuitBreakerService circuitBreakerService = new NoneCircuitBreakerService();
-
-        // Use the original Netty4Transport
+    private void startTransport() throws UnknownHostException {
         Netty4Transport transport = new Netty4Transport(
                 settings,
                 Version.CURRENT,
@@ -94,51 +78,13 @@ public class RunPlugin {
 
         transportService.start();
         transportService.acceptIncomingRequests();
+        final ActionListener actionListener = new ActionListener();
+        actionListener.runActionListener(true);
+    }
 
-        // Action Listener
-
-        boolean flag = true;
-        try (ServerSocket socket = new ServerSocket()) {
-            socket.bind(getLocalEphemeral(), 1);
-            socket.setReuseAddress(true);
-            DiscoveryNode dummy = new DiscoveryNode(
-                    "TEST",
-                    new TransportAddress(socket.getInetAddress(), socket.getLocalPort()),
-                    emptyMap(),
-                    emptySet(),
-                    version0
-            );
-            Thread t = new Thread() {
-                @Override
-                public void run() {
-                    try (Socket accept = socket.accept()) {
-                        if (flag) { // sometimes wait until the other side sends the message
-                            accept.getInputStream().read();
-                        }
-                    } catch (IOException e) {
-                        throw new UncheckedIOException(e);
-                    }
-                }
-            };
-            t.start();
-            ConnectionProfile.Builder builder = new ConnectionProfile.Builder();
-            builder.addConnections(
-                    1,
-                    TransportRequestOptions.Type.BULK,
-                    TransportRequestOptions.Type.PING,
-                    TransportRequestOptions.Type.RECOVERY,
-                    TransportRequestOptions.Type.REG,
-                    TransportRequestOptions.Type.STATE
-            );
-            builder.setHandshakeTimeout(TimeValue.timeValueHours(1));
-            //transportService.connectToNode(dummy, builder.build());
-            t.join();
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
+    public static void main(String[] args) throws UnknownHostException {
+        RunPlugin runPlugin = new RunPlugin();
+        runPlugin.startTransport();
     }
 
 }
